+++ src/test/java/nl/dtls/fairdatapoint/acceptance/search/query/List_GET.java
@@ -20,26 +20,23 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package nl.dtls.fairdatapoint.acceptance.search;
+package nl.dtls.fairdatapoint.acceptance.search.query;
 
 import nl.dtls.fairdatapoint.WebIntegrationTest;
-import nl.dtls.fairdatapoint.api.dto.search.SearchQueryDTO;
 import nl.dtls.fairdatapoint.api.dto.search.SearchQueryTemplateDTO;
-import nl.dtls.fairdatapoint.api.dto.search.SearchResultDTO;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.core.ParameterizedTypeReference;
 import org.springframework.http.*;
 
 import java.net.URI;
-import java.util.List;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.Is.is;
 import static org.hamcrest.core.IsEqual.equalTo;
 
 @DisplayName("GET /search/query")
-public class Query_GET extends WebIntegrationTest {
+public class List_GET extends WebIntegrationTest {
 
     private URI url() {
         return URI.create("/search/query");